add master layout, restructure folder
