+++ src/pages/Home.jsx
@@ -1,16 +1,14 @@
 import React from "react";
-import Header from "../components/header/Header";
 import HomePage from "../components/homePage/HomePage";
-import SearchBar from "../components/home/SearchBar";
-import Footer from "../components/footer/Footer";
+import SearchBar from "../components/homePage/SearchBar";
+import HeaderComponent from "../components/layout/Header";
+import FooterComponent from "../components/layout/Footer";
 
 function Home() {
   return (
     <div className="homePage">
-      <Header />
       <SearchBar />
       <HomePage />
-      <Footer />
     </div>
   );
 }
